+++ tests/test_laws.py
@@ -1,7 +1,3 @@
-import sys
-
-import pytest
-
 from returns.context import (
     Reader,
     ReaderFutureResult,
@@ -13,11 +9,6 @@ from returns.future import Future, FutureResult
 from returns.io import IO, IOResult
 from returns.maybe import Maybe
 from returns.result import Result
-
-pytestmark = pytest.mark.skipif(
-    sys.version_info < (3, 7),
-    reason='Python 3.6 does not support many hypothesis features',
-)
 
 check_all_laws(Maybe)
 check_all_laws(Result)